Creating session manager hook
